introduce browse-kill-ring-progn
This is a handy utility for defining keybindings. Keybindings have to be
`interactive`. They also don't have to be defined using `defun`; there's
no benefit to not using a lambda, since unlike e.g. hook functions, you never
need to identify the exact same function (e.g. to remove it).