introduce browse-kill-ring-quit-if-prefix-arg
Most of the prebuilt combination commands respect prefix-arg to mean
"quit after this operation". This is a combinator that helps us
implement this in other operations. It looks at the current prefix
argument given to whatever interactive operation was originally invoked,
without having to explicitly access it or pass it around.